FlightImproved
Bug on disabling gravity, when i tried to give input it wasn't taking it, i then instead of disabling gravity, added an aantigravity force that too wasn't working - realized that it was down direction i was adding force.
as for why movement so weird in flight can'tmove it in direction of joystick - reason no drag in air, once object reaces max velocity in one direction, it stays that direction foreevver, and doesn't register new inputs.
added rigidbody.damp